+++ x509/utils.go
@@ -1,6 +1,11 @@
 package x509
 
 import (
+	"bytes"
+	"crypto"
+	"crypto/rand"
+	"crypto/rsa"
+	"encoding/asn1"
 	"encoding/pem"
 	"errors"
 	"github.com/Hyperledger-TWGC/tjfoc-gm/sm2"
@@ -8,9 +13,12 @@ import (
 	"os"
 )
 
-func ReadPrivateKeyFromMem(data []byte, pwd []byte) (*sm2.PrivateKey, error) {
+func ReadPrivateKeyFromPem(FileName string, pwd []byte) (*sm2.PrivateKey, error) {
+	data, err := ioutil.ReadFile(FileName)
+	if err != nil {
+		return nil, err
+	}
 	var block *pem.Block
-
 	block, _ = pem.Decode(data)
 	if block == nil {
 		return nil, errors.New("failed to decode private key")
@@ -19,20 +27,11 @@ func ReadPrivateKeyFromMem(data []byte, pwd []byte) (*sm2.PrivateKey, error) {
 	return priv, err
 }
 
-func ReadPrivateKeyFromPem(FileName string, pwd []byte) (*sm2.PrivateKey, error) {
-	data, err := ioutil.ReadFile(FileName)
-	if err != nil {
-		return nil, err
-	}
-	return ReadPrivateKeyFromMem(data, pwd)
-}
-
-func WritePrivateKeytoMem(key *sm2.PrivateKey, pwd []byte) ([]byte, error) {
+func WritePrivateKeytoPem(FileName string, key *sm2.PrivateKey, pwd []byte) (err error) {
 	var block *pem.Block
-
 	der, err := MarshalSm2PrivateKey(key, pwd)
 	if err != nil {
-		return nil, err
+		return err
 	}
 	if pwd != nil {
 		block = &pem.Block{
@@ -45,15 +44,7 @@ func WritePrivateKeytoMem(key *sm2.PrivateKey, pwd []byte) ([]byte, error) {
 			Bytes: der,
 		}
 	}
-	return pem.EncodeToMemory(block), nil
-}
-
-func WritePrivateKeytoPem(FileName string, key *sm2.PrivateKey, pwd []byte) (err error) {
-	certPem, err := WritePrivateKeytoMem(key, pwd)
-	if err != nil {
-		return err
-	}
-
+	certPem := pem.EncodeToMemory(block)
 	file, err := os.Create(FileName)
 	if err != nil {
 		return err
@@ -68,41 +59,28 @@ func WritePrivateKeytoPem(FileName string, key *sm2.PrivateKey, pwd []byte) (err
 	return nil
 }
 
-func ReadPublicKeyFromMem(data []byte) (*sm2.PublicKey, error) {
-	block, _ := pem.Decode(data)
-	if block == nil || block.Type != "PUBLIC KEY" {
-		return nil, errors.New("failed to decode public key")
-	}
-	pub, err := ParseSm2PublicKey(block.Bytes)
-	return pub, err
-}
-
 func ReadPublicKeyFromPem(FileName string) (*sm2.PublicKey, error) {
 	data, err := ioutil.ReadFile(FileName)
 	if err != nil {
 		return nil, err
 	}
-	return ReadPublicKeyFromMem(data)
+	block, _ := pem.Decode(data)
+	if block == nil || block.Type != "PUBLIC KEY" {
+		return nil, errors.New("failed to decode public key")
+	}
+	return ParseSm2PublicKey(block.Bytes)
 }
 
-func WritePublicKeytoMem(key *sm2.PublicKey) ([]byte, error) {
+func WritePublicKeytoPem(FileName string, key *sm2.PublicKey) (err error) {
 	der, err := MarshalSm2PublicKey(key)
 	if err != nil {
-		return nil, err
+		return err
 	}
 	block := &pem.Block{
 		Type:  "PUBLIC KEY",
 		Bytes: der,
 	}
-	return pem.EncodeToMemory(block), nil
-}
-
-func WritePublicKeytoPem(FileName string, key *sm2.PublicKey) (err error) {
-	certPem, err := WritePublicKeytoMem(key)
-	if err != nil {
-		return err
-	}
-
+	certPem := pem.EncodeToMemory(block)
 	file, err := os.Create(FileName)
 	defer func() {
 		err = file.Close()
@@ -110,6 +88,158 @@ func WritePublicKeytoPem(FileName string, key *sm2.PublicKey) (err error) {
 	if err != nil {
 		return err
 	}
+	_, err = file.Write(certPem)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func ReadCertificateRequestFromPem(FileName string) (*CertificateRequest, error) {
+	data, err := ioutil.ReadFile(FileName)
+	if err != nil {
+		return nil, err
+	}
+	block, _ := pem.Decode(data)
+	if block == nil {
+		return nil, errors.New("failed to decode certificate request")
+	}
+	return ParseCertificateRequest(block.Bytes)
+}
+
+func CreateCertificateRequestToPem(FileName string, template *CertificateRequest, privKey *sm2.PrivateKey) error {
+	der, err := CreateCertificateRequest(rand.Reader, template, privKey)
+	if err != nil {
+		return err
+	}
+	block := &pem.Block{
+		Type:  "CERTIFICATE REQUEST",
+		Bytes: der,
+	}
+	file, err := os.Create(FileName)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		err = file.Close()
+	}()
+	err = pem.Encode(file, block)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func ReadCertificateFromPem(FileName string) (*Certificate, error) {
+	data, err := ioutil.ReadFile(FileName)
+	if err != nil {
+		return nil, err
+	}
+	block, _ := pem.Decode(data)
+	if block == nil {
+		return nil, errors.New("failed to decode certificate request")
+	}
+	return ParseCertificate(block.Bytes)
+}
+
+func CreateCertificateToPem(FileName string, template, parent *Certificate, pubKey *sm2.PublicKey, privKey *sm2.PrivateKey) error {
+	if template.SerialNumber == nil {
+		return errors.New("x509: no SerialNumber given")
+	}
+
+	hashFunc, signatureAlgorithm, err := signingParamsForPublicKey(privKey.Public(), template.SignatureAlgorithm)
+	if err != nil {
+		return err
+	}
+
+	publicKeyBytes, publicKeyAlgorithm, err := marshalPublicKey(pubKey)
+	if err != nil {
+		return err
+	}
+
+	asn1Issuer, err := subjectBytes(parent)
+	if err != nil {
+		return err
+	}
+
+	asn1Subject, err := subjectBytes(template)
+	if err != nil {
+		return err
+	}
+
+	if !bytes.Equal(asn1Issuer, asn1Subject) && len(parent.SubjectKeyId) > 0 {
+		template.AuthorityKeyId = parent.SubjectKeyId
+	}
+
+	extensions, err := buildExtensions(template)
+	if err != nil {
+		return err
+	}
+	encodedPublicKey := asn1.BitString{BitLength: len(publicKeyBytes) * 8, Bytes: publicKeyBytes}
+	c := tbsCertificate{
+		Version:            2,
+		SerialNumber:       template.SerialNumber,
+		SignatureAlgorithm: signatureAlgorithm,
+		Issuer:             asn1.RawValue{FullBytes: asn1Issuer},
+		Validity:           validity{template.NotBefore.UTC(), template.NotAfter.UTC()},
+		Subject:            asn1.RawValue{FullBytes: asn1Subject},
+		PublicKey:          publicKeyInfo{nil, publicKeyAlgorithm, encodedPublicKey},
+		Extensions:         extensions,
+	}
+
+	tbsCertContents, err := asn1.Marshal(c)
+	if err != nil {
+		return err
+	}
+
+	c.Raw = tbsCertContents
+
+	digest := tbsCertContents
+	switch template.SignatureAlgorithm {
+	case SM2WithSM3, SM2WithSHA1, SM2WithSHA256:
+		break
+	default:
+		h := hashFunc.New()
+		h.Write(tbsCertContents)
+		digest = h.Sum(nil)
+	}
+
+	var signerOpts crypto.SignerOpts
+	signerOpts = hashFunc
+	if template.SignatureAlgorithm != 0 && template.SignatureAlgorithm.isRSAPSS() {
+		signerOpts = &rsa.PSSOptions{
+			SaltLength: rsa.PSSSaltLengthEqualsHash,
+			Hash:       crypto.Hash(hashFunc),
+		}
+	}
+
+	var signature []byte
+	signature, err = privKey.Sign(rand.Reader, digest, signerOpts)
+	if err != nil {
+		return err
+	}
+	der, err := asn1.Marshal(certificate{
+		nil,
+		c,
+		signatureAlgorithm,
+		asn1.BitString{Bytes: signature, BitLength: len(signature) * 8},
+	})
+
+	if err != nil {
+		return err
+	}
+	block := &pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: der,
+	}
+	certPem := pem.EncodeToMemory(block)
+	file, err := os.Create(FileName)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		err = file.Close()
+	}()
 	_, err = file.Write(certPem)
 	if err != nil {
 		return err
